pyproject.toml
[tool.poetry]
name = "greendoor"
version = "0.1.0"
description = ""
authors = ["KIMHOONHEE <nmdkims@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
django-ninja = "0.16.1"
mysqlclient = "^2.1.0"
black = "21.12b0"
django-stubs = "1.9.0"
isort = "5.10.1"

mypy = "0.931"
bs4 = "^0.0.1"
requests = "^2.27.1"
PyMySQL = "^1.0.2"
boto3 = "^1.21.19"
django-storages = "^1.12.3"
Pillow = "^9.0.1"
django-crispy-forms = "^1.14.0"
django-allauth = "^0.49.0"
django-stubs-ext = "^0.3.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.crawling.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.isort]
profile = "black"




